requirements.txt
absl-py==2.1.0
apispec==6.7.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cors==1.0.1
cycler==0.12.1
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
flask-marshmallow==1.2.1
Flask-MySQL==1.6.0
Flask-Session==0.8.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.18
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgspec==0.18.6
mysql-connector-python==9.1.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.3.1
pipfile==0.0.2
protobuf==4.25.5
psutil==6.1.0
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.21.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==65.5.0
six==1.16.0
sounddevice==0.5.1
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tldextract==5.1.3
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.33
ultralytics-thop==2.0.11
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
zope.event==5.0
zope.interface==7.1.1
